[Inductor] Add fma.rn.f32.bf16 inline PTX for bf16 squaring (SM80+)
Add a post-grad pattern match that replaces pow(x_bf16, 2) with an
inline PTX fma.rn.f32.bf16 instruction, computing bf16×bf16→f32 in a
single instruction without explicit dtype conversion.

Requires codegen_upcast_to_fp32=False so loads stay in native bf16.
The pattern fires via the post-grad pass and emits
tl.inline_asm_elementwise with '=f,h,h' constraints.

Benchmark (GB200, RMSNorm bf16):
- [2048, 16384]: 1.13x speedup (53% → 60% SOL)
- [4096,  8192]: 1.06x speedup
- Neutral at large M (memory-bound)

Known limitation: codegen_upcast_to_fp32=False affects all ops in the
kernel, causing numerical differences for precision-sensitive ops like
rsqrt. A production version needs selective per-load upcast control.

+++ test/inductor/test_register_tiled_persistent_reduction.py
@@ -99,7 +99,7 @@ class TestRegisterTiledPersistentReduction(TestCase):
         """RMSNorm-like pattern."""
 
         def fn(x, weight):
-            variance = x.to(torch.float32).pow(2).mean(-1, keepdim=True)
+            variance = x.pow(2).mean(-1, keepdim=True)
             x_normed = x * torch.rsqrt(variance + 1e-6)
             return x_normed * weight
